+++ roxctl/scanner/generate/generate.go
@@ -1,9 +1,13 @@
 package generate
 
 import (
+	"fmt"
+	"strings"
+
 	"github.com/spf13/cobra"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/apiparams"
+	"github.com/stackrox/rox/pkg/istioutils"
 	"github.com/stackrox/rox/roxctl/common/util"
 	"github.com/stackrox/rox/roxctl/scanner/clustertype"
 	"github.com/stackrox/rox/roxctl/scanner/generate/run"
@@ -27,6 +31,10 @@ func Command() *cobra.Command {
 	c.Flags().BoolVar(&params.OfflineMode, "offline-mode", false, "whether to run the scanner in offline mode (so "+
 		"it doesn't reach out to the internet for updates)")
 	c.Flags().StringVar(&params.ScannerImage, "scanner-image", "", "Scanner image to use (leave blank to use server default)")
+	c.Flags().StringVar(&params.IstioVersion, "istio-support", "",
+		fmt.Sprintf(
+			"Generate deployment files supporting the given Istio version. Valid versions: %s",
+			strings.Join(istioutils.ListKnownIstioVersions(), ", ")))
 
 	return c
 }